Started sewing the network and elevator modules toghether Also organized the file structure a bit. Todo: make network and elevator driver run concurrently
