+++ Excercises/ex4/primary/primary.go
