+++ Project/driver-go-master/main.go
@@ -1,11 +1,10 @@
 package main
 
 import (
-	
+	"Driver-go/elevator"
 	"Driver-go/elevio"
 	"Driver-go/fsm"
 	"Driver-go/networkfunc"
-	"Driver-go/elevator"
 )
 
 func main() {
@@ -14,21 +13,7 @@ func main() {
 
 	elevio.Init("localhost:15657", numFloors)
 
-	conn :=	networkfunc.InitConn("localhost", "20008")
-	defer conn.Close()
-
-	var testelev elevator.Elevator
-	elevator.Elevator_uninitialized(&testelev)
 	
-	go networkfunc.WriteStateToUDP(conn, testelev)
-	go networkfunc.ReadFromUDP(conn)
-
-
-	select{
-	}
-	
-
-
 	drv_buttons := make(chan elevio.ButtonEvent)
 	drv_floors := make(chan int)
 	drv_obstr := make(chan bool)
@@ -39,10 +24,8 @@ func main() {
 	go elevio.PollObstructionSwitch(drv_obstr)
 	go elevio.PollStopButton(drv_stop)
 
-	
 
-
-	
 	fsm.Fsm_server(drv_buttons, drv_floors, drv_obstr, drv_stop)
+
 	
 }
