+++ Project/go.mod
@@ -1,0 +1,3 @@
+module Elev-project
+
+go 1.16
