--- driver-go-master/Network-go-master/go.mod
@@ -1,3 +1,0 @@
-module Network-go
-
-go 1.16
