--- driver-go-master/go.mod
@@ -1,3 +1,0 @@
-module Driver-go
-
-go 1.16
