--- networkMod/go.mod
@@ -1,3 +1,0 @@
-module ex4
-
-go 1.16
