--- networkMod/networkFunc/networkFunc.go
@@ -1,7 +1,0 @@
-package networkfunc
-
-//Initialize connection between master-slave
-func InitConn(){
-
-}
-
